# docs/swagger.yaml
basePath: /
definitions:
...
  api.confirmPurchaseRequest:
    properties:
      productId:
        type: string
      source:
        type: string
      verificationData:
        type: string
    type: object
...
  api.historyResponse:
    properties:
      challenges:
        items:
          $ref: '#/definitions/models.HistoryItem'
        type: array
      hasMore:
        description: |-
          HasMore says whether another page exists — what an infinite-scrolling
          client actually needs, so it never has to do the arithmetic.
        type: boolean
      limit:
        description: Limit is the window size actually applied; omitted when unpaginated.
        type: integer
      page:
        description: Page is 1-based; omitted when the caller asked for no page.
        type: integer
      total:
        description: Total is the row count across every page.
        type: integer
    type: object
  api.leaderboardResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/models.LeaderboardEntry'
        type: array
      me:
        $ref: '#/definitions/models.LeaderboardEntry'
    type: object
...
  api.pointsResponse:
    properties:
      points:
        type: integer
    type: object
  api.purchaseResponse:
    properties:
      expeditionStars:
        type: integer
      points:
        type: integer
    type: object
...
  api.spendPointsRequest:
    properties:
      amount:
        type: integer
    type: object
...
  httpx.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorDetail'
    type: object
  httpx.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
  models.Achievements:
    properties:
      firstPiece:
        type: boolean
      jigsawKing:
        type: boolean
      speedDemon:
        type: boolean
    type: object
...
  models.HistoryItem:
    properties:
      challengeId:
        type: string
      correctPieces:
        type: integer
      elapsedSeconds:
        type: integer
      finishedAt:
        type: string
      gridCols:
        type: integer
      gridRows:
        type: integer
      imageId:
        type: string
      imageUrl:
        type: string
      pieces:
        type: integer
      status:
        type: string
      thumbUrl:
        description: ThumbURL is [ImageURL] sized for the Recent Challenges strip.
        type: string
      title:
        type: string
    type: object
...
  models.LeaderboardEntry:
    properties:
      isMe:
        type: boolean
      level:
        type: integer
      name:
        type: string
      rank:
        type: integer
      score:
        type: integer
      userId:
        type: string
    type: object
...
  models.UserStats:
    properties:
      achievements:
        $ref: '#/definitions/models.Achievements'
      completionRate:
        type: integer
      highestScore:
        description: total correct pieces (cumulative)
        type: integer
      level:
        type: integer
      puzzlesSolved:
        type: integer
      rank:
        type: integer
      score:
        description: total correct pieces (ranking)
        type: integer
      solvedThisWeek:
        type: integer
      title:
        type: string
      totalTimeSeconds:
        type: integer
      xp:
        type: integer
      xpForNextLevel:
        type: integer
      xpIntoLevel:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Backend for the Jigsaw Puzzle app: email/password auth, image
    collection, admin daily challenges (Cloudinary) and per-day
    reputation ("Danh vọng").
  title: Jigsaw Puzzle API
  version: "1.0"
paths:
...
  /api/v1/leaderboard:
    get:
      description: Top players by score (total correctly-placed pieces) plus the caller's
        rank.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.leaderboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Global ranking
      tags:
      - me
...
  /api/v1/me/history:
    get:
      description: Recently played challenges (for the "Recent Challenges" list).
      parameters:
      - description: Page number (1-based); needs limit or defaults to 8
        in: query
        name: page
        type: integer
      - description: Rows per page; omit for the whole list
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.historyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: My challenge history
      tags:
      - me
...
  /api/v1/me/points:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.pointsResponse'
      security:
      - BearerAuth: []
      summary: Unlock points balance
      tags:
      - me
  /api/v1/me/points/purchase:
    post:
      consumes:
      - application/json
      parameters:
      - description: Store product id, source and verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.confirmPurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.purchaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Confirm an in-app purchase and credit its reward
      tags:
      - me
  /api/v1/me/points/spend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Amount to spend (default 1)
        in: body
        name: request
        schema:
          $ref: '#/definitions/api.spendPointsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.pointsResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Spend unlock points
      tags:
      - me
...
  /api/v1/me/stats:
    get:
      description: Level, XP, score, rank, completion rate and achievements for the
        current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserStats'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: My stats
      tags:
      - me
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token issued by
      /api/v1/auth/login or /register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"